+++ manager.go
@@ -30,22 +30,22 @@ func getManagerFromLookup(u *ldap.Entry, buffer *bytes.Buffer) string {
 		return ""
 	}
 	if ldapImportConf.UserManagerMapping.GetIDFromName {
+		buffer.WriteString(loggerGen(1, "LDAP Manager String: "+ManagerAttributeName))
 		ManagerAttributeName = getNameFromLDAPString(ManagerAttributeName, buffer)
 	}
-	buffer.WriteString(loggerGen(1, "Looking Up Manager "+ManagerAttributeName))
+	buffer.WriteString(loggerGen(1, "Looking Up Manager from Cache: "+ManagerAttributeName))
 	managerIsInCache, ManagerIDCache := managerInCache(ManagerAttributeName)
 
 	//-- Check if we have Chached the site already
 	if managerIsInCache {
-		buffer.WriteString(loggerGen(1, "Found Manager in Cache "+ManagerIDCache))
+		buffer.WriteString(loggerGen(1, "Found Manager in Cache: "+ManagerIDCache))
 		return ManagerIDCache
 	}
-	buffer.WriteString(loggerGen(1, "Manager Not In Cache Searching"))
+	buffer.WriteString(loggerGen(1, "Manager Not In Cache Searching Hornbill"))
 	ManagerIsOnInstance, ManagerIDInstance := searchManager(ManagerAttributeName, buffer)
 	//-- If Returned set output
 	if ManagerIsOnInstance {
-		buffer.WriteString(loggerGen(1, "Manager Lookup found Id "+ManagerIDInstance))
-
+		buffer.WriteString(loggerGen(1, "Manager Lookup found Id: "+ManagerIDInstance))
 		return ManagerIDInstance
 	}