+++ user.go
@@ -108,7 +108,7 @@ func updateUser(u *ldap.Entry, buffer *bytes.Buffer, espXmlmc *apiLib.XmlmcInstS
 			userAddRoles(userID, buffer, espXmlmc)
 		}
 		//-- Process Profile Details
-		boolUpdateProfile := userUpdateProfile(u, buffer, espXmlmc)
+		boolUpdateProfile := userUpdateProfile(u, buffer, espXmlmc, "Update")
 		if boolUpdateProfile != true {
 			err = errors.New("Error Updating User Profile")
 			errorCountInc()
@@ -124,7 +124,7 @@ func updateUser(u *ldap.Entry, buffer *bytes.Buffer, espXmlmc *apiLib.XmlmcInstS
 		return true, nil
 	}
 	//-- Process Profile Details as part of the dry run for testing
-	boolUpdateProfile := userUpdateProfile(u, buffer, espXmlmc)
+	boolUpdateProfile := userUpdateProfile(u, buffer, espXmlmc, "Update")
 	if boolUpdateProfile != true {
 		err := errors.New("Error Updating User Profile")
 		errorCountInc()
@@ -218,7 +218,7 @@ func createUser(u *ldap.Entry, buffer *bytes.Buffer, espXmlmc *apiLib.XmlmcInstS
 			userAddRoles(userID, buffer, espXmlmc)
 		}
 		//-- Process Profile Details
-		boolUpdateProfile := userUpdateProfile(u, buffer, espXmlmc)
+		boolUpdateProfile := userUpdateProfile(u, buffer, espXmlmc, "Create")
 		if boolUpdateProfile != true {
 			err = errors.New("Error Updating User Profile")
 			errorCountInc()
@@ -229,7 +229,7 @@ func createUser(u *ldap.Entry, buffer *bytes.Buffer, espXmlmc *apiLib.XmlmcInstS
 		return true, nil
 	}
 	//-- Process Profile Details as part of the dry run for testing
-	boolUpdateProfile := userUpdateProfile(u, buffer, espXmlmc)
+	boolUpdateProfile := userUpdateProfile(u, buffer, espXmlmc, "Create")
 	if boolUpdateProfile != true {
 		err := errors.New("Error Updating User Profile")
 		errorCountInc()